+++ examples/multiProcessExample/src/main/java/io/realm/examples/realmmultiprocessexample/MyApplication.java
@@ -1,0 +1,16 @@
+package io.realm.examples.realmmultiprocessexample;
+
+import android.app.Application;
+
+import io.realm.Realm;
+import io.realm.RealmConfiguration;
+
+public class MyApplication extends Application {
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        Realm.init(this);
+        RealmConfiguration configuration = new RealmConfiguration.Builder().deleteRealmIfMigrationNeeded().build();
+        Realm.setDefaultConfiguration(configuration);
+    }
+}