+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Enhancements
 
+* Added support for using non-encrypted Realm in multiple processes from the same apk (#1091).
+
 ### Bug Fixes
 
 ### Interal